backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
huggingface-hub
python-multipart